Integrate with contracts repo
Fix #1
