+++ .gitignore
@@ -1,5 +1,6 @@
 .vscode
 .DS_Store
+src/contracts
 
 # Logs
 logs
