+++ tsconfig.json
@@ -1,6 +1,8 @@
 {
   "extends": "@tsconfig/svelte/tsconfig.json",
-
   "include": ["src/**/*"],
+  "compilerOptions": {
+    "resolveJsonModule": true
+  },
   "exclude": ["node_modules/*", "__sapper__/*", "public/*"]
 }
